Removed old DynamicFPSLimiter v3.0.0 Folder and files (since this was manually added) Renamed DFL_v3.0.0.py to DFL_v3.py added version.txt to Source folder
