--- DynamicFPSLimiter_v3.0.0/profiles.ini
@@ -1,11 +1,0 @@
-[Global]
-maxcap = 60
-mincap = 30
-capstep = 5
-usagecutofffordecrease = 80
-delaybeforedecrease = 2
-usagecutoffforincrease = 70
-delaybeforeincrease = 2
-minvalidgpu = 20
-minvalidfps = 20
-
