--- DynamicFPSLimiter_v3.0.0/settings.ini
@@ -1,3 +1,0 @@
-[Preferences]
-ShowPlot = True
-ShowTooltip = True
