+++ Source/git_commands.txt
@@ -1,4 +1,4 @@
-# to make changes and add it to the github
+# To make changes and add it to the github
 
 git add .
 git commit -m "Message"
