+++ Source/version.txt
@@ -1,0 +1,26 @@
+VSVersionInfo:
+    ffi=FixedFileInfo(
+        filevers=(3, 0, 1, 0),
+        prodvers=(3, 0, 1, 0),
+        mask=0x3f,
+        flags=0x0,
+        OS=0x40004,
+        fileType=0x1,
+        subtype=0x0,
+        date=(0, 0)
+    ),
+    kids=[
+        StringFileInfo(
+            [
+            StringTable(
+                u'040904B0',
+                [StringStruct(u'CompanyName', u'SameSalamander5710'),
+                StringStruct(u'FileDescription', u'DynamicFPSLimiter'),
+                StringStruct(u'FileVersion', u'3.0.1.0'),
+                StringStruct(u'InternalName', u'DynamicFPSLimiter'),
+                StringStruct(u'OriginalFilename', u'DynamicFPSLimiter.exe'),
+                StringStruct(u'ProductName', u'DynamicFPSLimiter'),
+                StringStruct(u'ProductVersion', u'3.0.1.0')])
+            ]),
+        VarFileInfo([VarStruct(u'Translation', [1033, 1200])])
+    ]
